requirements.txt
apistar==0.5.42
apistar-crud==0.6.5
apistar-pagination==0.3.0
certifi==2018.8.24
chardet==3.0.4
Click==7.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.2
django-appconf==1.0.2
django-bootstrap-form==3.4
django-crispy-forms==1.7.2
django-extensions==2.1.0
django-select2==6.0.2
django-sortable-column==0.2.2
django-toolbelt==0.0.1
djangoajax==3.0.2
djangorestframework==3.7.7
easy-thumbnails==2.5
gunicorn==19.9.0
idna==2.7
Jinja2==2.10
MarkupSafe==1.0
Pillow==5.3.0
psycopg2==2.7.5
pytz==2018.5
PyYAML==3.13
requests==2.19.1
six==1.11.0
static3==0.7.0
urllib3==1.23
virtualenv==16.0.0
Werkzeug==0.14.1
whitenoise==4.1
